refactor: move _coerce_bool to index/utils as the shared home
Both indexing paths and the CLI now use it; page_index_md re-exports it
for existing importers.
